Enhance theming support by adding new aliases for settings and theme paths in Babel config. Refactor BaniList and other components to remove unnecessary props and streamline imports. Remove legacy styles and unused theme definitions for improved maintainability.
