+++ src/AboutScreen/hooks/useHeader.js
@@ -1,9 +1,8 @@
+import React, { useEffect } from "react";
+import { Icon } from "@rneui/themed";
+import useTheme from "@common/context";
 import { STRINGS } from "@common";
 import useThemedStyles from "@common/hooks/useThemedStyles";
-import { Icon } from "@rneui/themed";
-import React, { useEffect } from "react";
-import useTheme from "@common/context";
-
 import createStyles from "../styles";
 
 const useHeader = (navigation) => {
