+++ src/common/hooks/useThemedStyles.js
@@ -5,5 +5,5 @@ import { useTheme } from "../context/ThemeContext";
 export default function useThemedStyles(create) {
   const { theme } = useTheme();
   // create(theme) should return a plain object of style rules
-  return useMemo(() => StyleSheet.create(create(theme)), [theme, create]);
+  return useMemo(() => StyleSheet.create(create(theme)), [theme]);
 }
