+++ src/theme/index.js
@@ -1,0 +1,2 @@
+export { default as lightTheme } from "./lightTheme";
+export { default as darkTheme } from "./darkTheme";
