+++ src/theme/spacing.js
@@ -1,0 +1,12 @@
+const spacing = {
+  xs: 2,
+  sm: 4,
+  md: 8,
+  lg: 16,
+  xl: 24,
+  xxl: 32,
+  xxxl: 48,
+  huge: 64,
+};
+
+export default spacing;
