+++ src/theme/staticColors.js
@@ -1,0 +1,7 @@
+const staticColors = {
+  WHITE_COLOR: "#faf9f6", // used on Home Screen Header Ik ongkar and title
+  NIGHT_BLACK: "#121212",
+  NIGHT_OPACITY_BLACK: "rgba(0, 0, 0, 0.5)",
+};
+
+export default staticColors;
